requirements.txt
# ── Web app ───────────────────────────────────────────────────────────
streamlit>=1.35.0
streamlit-option-menu>=0.3.6

# ── Data ──────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0
openpyxl>=3.1.0          # Excel read/write
xlrd>=2.0.1              # Legacy .xls support

# ── Machine learning ──────────────────────────────────────────────────
scikit-learn>=1.4.0

# ── Visualisation ─────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Config ────────────────────────────────────────────────────────────
pyyaml>=6.0.1            # Load config/models.yaml

# ── Dev / testing ─────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0

# ── Explainability ────────────────────────────────────────────────────
shap>=0.45.0
statsmodels>=0.14.0      # LOWESS smoothing for SHAP dependence + PDP

# ── Optimisation ──────────────────────────────────────────────────────
scikit-optimize>=0.9.0   # Bayesian optimisation (skopt)
scipy>=1.12.0            # Dirichlet sampling, stats

# ── ML Models (from code0) ────────────────────────────────────────────
catboost>=1.2.0          # CatBoost (optional — remove if not using)
plotly>=5.20.0           # Interactive charts for safe region maps
